pyproject.toml
[tool.poetry]
name = "agentic-research-assistant"
version = "0.1.0"
description = "An agentic research assistant powered by LangGraph and Claude"
authors = ["Research Assistant Team"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.scripts]
research-agent = "src.cli.main:app"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langgraph = ">=1.1.3,<2.0"
langchain-anthropic = ">=0.3.0"
anthropic = ">=0.40.0"
tavily-python = ">=0.5.0"
faiss-cpu = ">=1.8.0"
sentence-transformers = ">=3.0.0"
pydantic = ">=2.0.0"
fastapi = ">=0.115.0"
uvicorn = { extras = ["standard"], version = ">=0.30.0" }
structlog = ">=24.0.0"
python-dotenv = ">=1.0.0"
tenacity = ">=8.2.0"
typer = { extras = ["all"], version = ">=0.12.0" }
httpx = ">=0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
pytest-cov = ">=5.0.0"
ruff = ">=0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
